Fix build for Windows 10/11
- Change `iDotMatrix Controller.exe` filename to `iDotMatrixController.exe` to avoid spaces in binary filenames.
- Change `build.bat` so it will activate the venv, install build tooling and qt for gui then run pyinstaller against the gui and cli app. An additional `idotmatrix.exe` binary will be built to allow cli control from other frontends.
- Change `app.py` commmand in `gui.py` to use `idotmatrix.exe` cli binary.
- Change CMD class to move it into `app.py` as `pyinstaller` could not correctly build `app.py` with the CMD class in a subfolder.
- Remove `cmd.py` from the core folder as it is no longer needed
- Add `create_venv.bat` script for Windows so it will correctly create the Windows python venv and install project dependencies for `app.py`
