+++ build.bat
@@ -1,5 +1,1 @@
-pip install pyqt5
-
-pip install pyinstaller
-
-pyinstaller gui.py -F -w -n "iDotMatrix Controller" -i idmc.ico --distpath ./
+cmd /k "cd /d venv\Scripts & activate & cd ../../ & python -m pip install pyqt5 & python -m pip install pyinstaller & pyinstaller app.py -F -w -n idotmatrix -i idmc.ico --distpath ./dist & pyinstaller gui.py -F -w -n ""iDotMatrixController"" -i idmc.ico --distpath ./dist"
