+++ create_venv.bat
@@ -1,0 +1,2 @@
+python -m venv venv
+cmd /k "cd /d venv\Scripts & activate & cd ../../ & python -m pip install ."
