+++ gui.py
@@ -469,7 +469,7 @@ class PixelPaintDialog(QDialog):
 
     def run_command(self, command_array):
         process = QProcess(self)
-        process.start("cmd.exe", ["/c", "py", "app.py"] + command_array)
+        process.start("cmd.exe", ["/c", "%cd%/idotmatrix"] + command_array)
         process.waitForFinished()
 
     def clear_device(self):
@@ -658,7 +658,7 @@ class DevicePage(QWidget):
             self.process.setProcessChannelMode(QProcess.MergedChannels)
             self.process.readyRead.connect(self.handle_ready_read)
             self.process.finished.connect(self.process_finished)
-            self.process.start("cmd.exe", ["/c", "py", "app.py"] + args)
+            self.process.start("cmd.exe", ["/c", "%cd%/idotmatrix"] + args)
 
     def handle_ready_read(self):
         data = self.process.readAll()
@@ -996,7 +996,7 @@ class MainWindow(QWidget):
         self.process.setProcessChannelMode(QProcess.MergedChannels)
         self.process.readyRead.connect(self.handle_ready_read)
         self.process.finished.connect(self.process_finished)
-        self.process.start("cmd.exe", ["/c", "py app.py --scan"])
+        self.process.start("cmd.exe", ["/c", "%cd%/idotmatrix --scan"])
 
     def handle_ready_read(self):
         data = self.process.readAll()
